Refactor state management and fix refresh race condition
