+++ webapp/.gitignore
@@ -22,3 +22,7 @@ dist-ssr
 *.njsproj
 *.sln
 *.sw?
+
+# Coverage and lockfiles
+coverage/
+bun.lock
