+++ webapp/src/components/GdbComponents/BreakPoints/BreakPoints.jsx
@@ -1,8 +1,6 @@
 import React, { useEffect, useState } from "react";
 import { DataState } from "./../../../context/DataContext";
 import "./BreakPoints.css";
-import axios from "axios";
-
 const data = [
   {
     offset: "0x2fffa36f603112ffff34",
@@ -23,21 +21,8 @@ const data = [
 ];
 
 const BreakPoints = () => {
-  const { refresh, setInfoBreakpointData, infoBreakpointData } = DataState();
+  const { infoBreakpointData } = DataState();
 
-  const fetchInfoBreakpoints = async () => {
-    const data = await axios.post("http://127.0.0.1:10000/info_breakpoints", {
-      name: "program",
-    });
-    console.log(data.data["result"]);
-    setInfoBreakpointData(data.data["result"]);
-  };
-  useEffect(() => {
-    if (refresh) {
-      console.log("click from breakpoint in GdbComponents");
-      fetchInfoBreakpoints();
-    }
-  }, [refresh]);
   return (
     <div>
       {/* BreakPoints */}
@@ -45,7 +30,7 @@ const BreakPoints = () => {
         {infoBreakpointData
           ? infoBreakpointData
           : data?.length > 0
-          ? data.map((obj) => {
+            ? data.map((obj) => {
               return (
                 <div>
                   <div>{obj.offset}</div>
@@ -53,8 +38,8 @@ const BreakPoints = () => {
                 </div>
               );
             })
-          : infoBreakpointData}
-        {}
+            : infoBreakpointData}
+        { }
       </div>
     </div>
   );
