+++ webapp/src/components/GdbComponents/MemoryMap/MemoryMap.jsx
@@ -2,8 +2,6 @@ import React, { useEffect } from "react";
 import { DataState } from "./../../../context/DataContext";
 
 import "./MemoryMap.css";
-import axios from "axios";
-
 const data = [
   "0x7fffffffe270: 0x00 0x00 0x00 0x00 0x00 0x00 0x00 0x00",
   "0x7fffffffe270: 0x00 0x00 0x00 0x00 0x00 0x00 0x00 0x00",
@@ -14,21 +12,9 @@ const data = [
   "0x7fffffffe270: 0x00 0x00 0x00 0x00 0x00 0x00 0x00 0x00",
   "0x7fffffffe270: 0x00 0x00 0x00 0x00 0x00 0x00 0x00 0x00",
 ];
+
 const MemoryMap = () => {
-  const { refresh, memoryMap, setMemoryMap } = DataState();
-
-  const fetchMemoryMap = async () => {
-    console.log("Click form memory map");
-    const data = await axios.post("http://127.0.0.1:10000/memory_map", {
-      name: "program",
-    });
-    console.log(data.data.result);
-    setMemoryMap(data.data.result);
-  };
-
-  useEffect(() => {
-    if (refresh) fetchMemoryMap();
-  }, [refresh]);
+  const { memoryMap } = DataState();
 
   return (
     <div>
@@ -37,10 +23,10 @@ const MemoryMap = () => {
         {memoryMap
           ? memoryMap
           : data?.length > 0
-          ? data.map((obj) => {
+            ? data.map((obj) => {
               return <a>{obj}</a>;
             })
-          : ""}
+            : ""}
       </div>
     </div>
   );
